Create bagging.py
Modified version of the bagging.py from sklearn. The method generate_indices_stratified was added in order to force the bootstrapping to select subsets of data, that reflect the initial proportion of the classes. For example, if in the training data there is a ratio of 2:1 of the positive class versus the negative class and if we consider a sampling rate r, a bootstrap sample will contain 2r instances of positive class and r instances of the negative class.